🔧 refactor(ProjectAnalyticsDto): move ProjectAnalyticsDto to the dto package and update imports in related files
