+++ se347-backend/src/main/java/com/github/ngodat0103/se347_backend/dto/project/ProjectAnalyticsDto.java
@@ -1,4 +1,4 @@
-package com.github.ngodat0103.se347_backend.persistence.document.project;
+package com.github.ngodat0103.se347_backend.dto.project;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -12,6 +12,8 @@ public class ProjectAnalyticsDto {
   private int assignedTaskDifference;
   private int completedTaskCount;
   private int completedTaskDifference;
+  private int inCompletedTaskCount;
+  private int inCompletedTaskDifference;
   private int overdueTaskCount;
   private int overdueTaskDifference;
 }
