+++ se347-backend/src/main/java/com/github/ngodat0103/se347_backend/service/project/ProjectService.java
@@ -1,7 +1,7 @@
 package com.github.ngodat0103.se347_backend.service.project;
 
+import com.github.ngodat0103.se347_backend.dto.project.ProjectAnalyticsDto;
 import com.github.ngodat0103.se347_backend.dto.project.ProjectDto;
-import com.github.ngodat0103.se347_backend.persistence.document.project.ProjectAnalyticsDto;
 import java.io.InputStream;
 import java.util.Set;
 import org.springframework.http.MediaType;
